Move "API Core" to "API Client"
